requirements.txt
beautifulsoup4==4.12.2
google-genai==1.13.0
html-to-json==2.0.0
openai==1.76.2
python-dotenv==1.0.0 
requests==2.31.0
